+++ index.ts
@@ -1,1 +1,0 @@
-export { babelInlineScripts as default } from "./src";