+++ src/constants/script-tag-regexp.ts
@@ -1,0 +1,1 @@
+export default /<script.*?>([\s\S]*?)<\/script>/gim;