+++ src/utils/extract-script-tags.ts
@@ -1,0 +1,3 @@
+import { scriptTagRegexp } from "../constants/index.js";
+
+export default (content: string) => content.match(scriptTagRegexp);