+++ src/utils/strip-script-tags.ts
@@ -1,0 +1,4 @@
+import { scriptTagRegexp } from "../constants/index.js";
+
+export default (htmlString: string) =>
+  htmlString.replace(scriptTagRegexp, "$1");