+++ src/utils/write-file-to-url.ts
@@ -1,0 +1,8 @@
+import { writeFileSync } from "fs";
+import { fileURLToPath } from "url";
+
+export default (fileContent: string, fileURL: string) => {
+  const HTMLPath = fileURLToPath(fileURL);
+  const byteContent = Buffer.from(fileContent);
+  writeFileSync(HTMLPath, byteContent);
+};